Factor out captured variables that are not reused outside findGit callback
